ANTILAG: Split weapon definitions into weaponpred_defs.c
Move the wpredict_definitions table, WeaponDefinition_SendEntity and
WPredict_Initialize out of antilag.c into a dedicated weaponpred_defs.c.
Pure code motion; both functions are already declared in g_local.h and
the table has no external references.
